Moves request caching out of Store
Adds an explicit cache at the loopbackGraphql layer that is initialized
with server-side requests. Removes akward store-population from
ReportLayout and store reading from getInitialProps.

Moves RequestForm to be localized at the RequestDialog component so that
its lifetime naturally matches that component. Pushes more things around
as props rather than hanging off of the store so that dependencies are
more obvious.